requirements.txt
numpy
pandas
openpyxl
xlrd
selenium==3.7.0
beautifulsoup4
urllib3==1.22
chromedriver-binary
Django==2.0.1
djongo==1.2.19
dj-static==0.0.6
djangorestframework==3.7.7
django-cors-headers==2.1.0
gunicorn==19.7.1
google-api-core==0.1.1
google-auth==1.2.0
google-cloud-core==0.28.0
google-cloud-vision==0.28.0
google-gax
googleapis-common-protos==1.5.3
